requirements.txt
dvc==3.51.2
dvc-gdrive==3.0.1
fastapi==0.111.1
imbalanced-learn==0.12.3
joblib==1.4.2
mlflow==2.14.3
mkdocs==1.6.0
mkdocs-material==9.5.31
numpy ==1.26.4
pandas==2.2.2
pytest==8.3.2
PyYAML==6.0.1
scikit-learn==1.5.1
setuptools==70.3.0
uvicorn==0.30.1
tenacity>=8.4.0
jsonpatch<=2.0
requests-toolbelt<=2.0.0
zstandard<=0.24.0
pydantic-settings<=3.0.0
seaborn>=0.13.1
mlflow>=2.14.2
evidently>=0.6.6
-e .
# evidently @ git+https://github.com/evidentlyai/evidently.git